Replace incorrect board name in list of supported boards
The readme specifies the project's target of ATmega328P-based boards and includes a list of examples of such boards.

Previously that list included the "Micro", but this board is based on the ATmega32U4 microcontroller rather than
ATmega328P. From experience, I know that people often mix up the "Micro"/"Pro Micro" with the similarly named and formed
"Pro Mini" board, which is a common ATmega328P-based board. So I replaced the incorrect mention of "Micro" with "Pro
Mini".