Added an option HOMING_SET_ORIGIN_AFTER_PULLOFF
This option sets the machine position to zero (or max_travel if homing
direction is inverted) after instead of before performing the homing
pulloff move. When soft limits are enabled, this provides a virtual
buffer zone between the machine's 0 and the hardware limit switch that
prevents the operator being able to jog right up to the switch and
potentially accidentally trigger a hard limit alarm.